requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
prophet==1.1.4
kafka-python==2.0.2
paho-mqtt==1.6.1
redis==5.0.1
scipy==1.11.4
joblib==1.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
aiofiles==23.2.1
websockets==12.0